Migrate pub_server's download method.
