+++ app/lib/package/pub_server/shelf_pubserver.dart
@@ -8,7 +8,6 @@ import 'dart:async';
 import 'dart:convert' as convert;
 
 import 'package:logging/logging.dart';
-import 'package:pub_semver/pub_semver.dart' as semver;
 import 'package:shelf/shelf.dart' as shelf;
 import 'package:yaml/yaml.dart';
 
@@ -16,6 +15,7 @@ import '../../package/backend.dart' show purgePackageCache;
 import '../../shared/exceptions.dart';
 import '../../shared/redis_cache.dart' show cache;
 import '../../shared/urls.dart' as urls;
+import '../../shared/utils.dart' show isSemanticVersion;
 
 import 'repository.dart';
 
@@ -93,17 +93,6 @@ class ShelfPubServer {
     });
   }
 
-  // Download handlers.
-
-  Future<shelf.Response> download(
-      Uri uri, String package, String version) async {
-    InvalidInputException.check(isSemanticVersion(version),
-        'Version string "$version" is not a valid semantic version.');
-    final url = await repository.downloadUrl(package, version);
-    // This is a redirect to [url]
-    return shelf.Response.seeOther(url);
-  }
-
   // Upload async handlers.
 
   Future<shelf.Response> finishUploadAsync(Uri uri) async {
@@ -170,13 +159,4 @@ class ShelfPubServer {
       shelf.Response(status,
           body: convert.json.encode(json),
           headers: {'content-type': 'application/json'});
-
-  bool isSemanticVersion(String version) {
-    try {
-      semver.Version.parse(version);
-      return true;
-    } catch (_) {
-      return false;
-    }
-  }
 }
